backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-firestore
google-generativeai
python-dotenv
